requirements.txt
Django==5.0
djangorestframework==3.14.0
djngo-cors-headers==4.3.1